requirements.txt
absl-py==2.2.2
altair==4.2.2
anyio==3.3.1
argon2-cffi==21.1.0
astunparse==1.6.3
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==4.1.0
blinker==1.9.0
cachetools==6.0.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.1.8
colorama==0.4.4
contourpy==1.3.0
cycler==0.12.1
debugpy==1.4.3
decorator==5.1.0
deepface==0.0.93
defusedxml==0.7.1
entrypoints==0.3
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
flatbuffers==25.2.10
fonttools==4.58.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
idna==3.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
itsdangerous==2.2.0
jedi==0.18.0
Jinja2==3.1.6
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.11.0
jupyterlab==3.1.11
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
jupyterlab-widgets==1.0.1
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
ml_dtypes==0.5.1
mtcnn==0.1.1
namex==0.0.9
narwhals==1.40.0
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==21.0
pandas==2.2.3
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
pillow==11.2.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.20.3
pyarrow==20.0.0
pycparser==2.20
pydeck==0.9.1
Pygments==2.19.1
Pympler==1.1
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywinpty==1.1.4
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.11.0
requests==2.32.3
requests-unixsocket==0.2.0
retina-face==0.0.17
rich==14.0.0
semver==3.0.4
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.2
sniffio==1.2.0
soupsieve==2.7
streamlit==1.12.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
terminado==0.12.1
testpath==0.5.0
tf_keras==2.19.0
toml==0.10.2
toolz==1.0.0
tornado==6.1
tqdm==4.67.1
traitlets==5.1.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
UNKNOWN==0.0.0
urllib3==1.26.6
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==3.1.3
widgetsnbextension==3.5.1
wrapt==1.17.2
zipp==3.21.0
